% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step}
\alias{step}
\title{Create a .R file and template}
\usage{
step(name, ext = "R")
}
\arguments{
\item{name}{File name, without extension; will create if it doesn't already exist.}

\item{ext}{An optional extension for the template file. Defaults to `R`}
}
\description{
Create a .R file and template
}
